Reestructure frontend and add Routes
