+++ my-code/whats-in/src/pages/Homepage/index.jsx
@@ -1,0 +1,5 @@
+export default function Homepage() {
+    return (
+        <p>Homepage</p>
+    )
+}
