+++ my-code/whats-in/src/pages/Movie/index.jsx
@@ -1,0 +1,5 @@
+export default function Moviepage() {
+    return(
+        <p>Movie page</p>
+    )
+}
